requirements.txt
setuptools~=75.8.0
ollama~=0.4.7
scikit-learn~=1.6.1
matplotlib~=3.10.0
seaborn~=0.13.2
PyYAML~=6.0.2
reportlab~=4.3.0
python-docx~=1.1.2